Refactoring of QStringPropertyEditor to use NSFormatters
Refactored of QStringPropertyEditor to use NSFormatters. This allow the property editor to be used for non-string properties where the user still types the value as text.

QStringPropertyEditor is now QTextInputPropertyEditor. QStringPropertyEditor is simply now a typedef to QTextInputPropertyEditor. QTextInputPropertyEditor can optionally be initialized with a NSFormatter that is used to convert between the property value and it's text representation.

Added QNumberPropertyEditor for editing numeric properties.