Added new editing state (QTextEditingModeFinishingForced) to handle case when a text field loses first responder status
The a text field is forced to lose first responder status, we want the edits to be committed to the model object if validation passes. This is needed in this common UI scenario: a form is contained in a navigation controller and the user edits a text field and presses the back button. Prior to this commit, the edits would be lost. Now, if validation passes, the changes will be committed.