Improved display and animation of validation error messages in QTextFieldTableViewCell and QTextInputPropertyEditor
Previously, when a validation message was displayed, it would appear immediately at the start of the table cell animation. This would cause the positioning of the error message label to animated (which looked bad), and the text to overlap the cell below. Now we take care to show the label after the table cell animation is complete, thus ensuring that the label doesn't overlap any other cells, and there's no visible animation of the label's position.