Reproduce iOS default table view cell heights
The custom table view cells use auto layout and are self-sizing. However, this produces some inconsistencies relative to iOS defaults. In particular, the following traits are seen in table view cells used by native iOS apps:

- Even at font sizes smaller than the default (Large), the table cell height is never smaller than 44 points.
- Table cell heights are larger than a label’s intrinsic height plus system standard spacing above and below.

To reproduce this, a ≥ height constraint is added to the table cell content view that mimics the above traits.