Allow properties from different objects to be edited
Each property editor now references the object that it edits. Previously, only RSObjectEditorViewController maintained a reference to the edited object, and passed this (or the property value) to the property editor when it needed.

Each property editor now functions as an independent controller, mediating between the model object property value and its UI (a table view cell).