Remove requirement of property editors to update the modified state of a form
Form modifications are now handled by RPropertyFormItem, which observes the property on the model; if the value changes, it updates the modified state.

(This slightly changes the semantics of the modified state; previously a form would be modified only when the model object was modified via the form. Now the form will be considered modified even when the model object is externally modified.)